+++ src/greynoise/api/__init__.py
@@ -12,7 +12,8 @@ from greynoise.__version__ import __version__
 from greynoise.api.analyzer import Analyzer
 from greynoise.api.filter import Filter
 from greynoise.exceptions import RateLimitError, RequestFailure
-from greynoise.util import load_config, validate_ip
+from greynoise.util import load_config, validate_ip, validate_timeline_field_value, \
+    validate_timeline_days, validate_timeline_granularity
 
 LOGGER = logging.getLogger(__name__)
 
@@ -45,6 +46,8 @@ class GreyNoise(object):  # pylint: disable=R0205,R0902
     EP_NOISE_CONTEXT_MULTI = "v2/noise/multi/context"
     EP_COMMUNITY_IP = "v3/community/{ip_address}"
     EP_SIMILARITY_IP = "v3/similarity/ips/{ip_address}"
+    EP_TIMELINE_IP = "v3/noise/ips/{ip_address}/timeline"
+    EP_TIMELINE_DETAILS_IP = "v3/noise/ips/{ip_address}/hourly-summary"
     EP_META_METADATA = "v2/meta/metadata"
     EP_PING = "ping"
     EP_RIOT = "v2/riot/{ip_address}"
@@ -163,7 +166,6 @@ class GreyNoise(object):  # pylint: disable=R0205,R0902
         }
 
         url = "/".join([self.api_server, endpoint])
-
 
         LOGGER.debug("Sending API request...URL: %s", url)
         LOGGER.debug("Sending API request...method: %s", method)
@@ -566,13 +568,16 @@ class GreyNoise(object):  # pylint: disable=R0205,R0902
 
         return response
 
-    def similar(self, ip_address):
+    def similar(self, ip_address, limit=None):
         """Query IP on the IP Similarity API
 
         :param ip_address: IP address to use in the look-up.
         :type ip_address: str
+        :param limit: Limit the number of matches returned by the endpoint
+        :type limit: str
         :return: Context for the IP address.
         :rtype: dict
+
 
         """
         if self.offering == "community":
@@ -581,7 +586,91 @@ class GreyNoise(object):  # pylint: disable=R0205,R0902
             LOGGER.debug("Checking IP Sim results for %s...", ip_address)
             validate_ip(ip_address)
 
-            endpoint = self.EP_SIMILARITY_IP.format(ip_address=ip_address)
+            if limit is None:
+                endpoint = self.EP_SIMILARITY_IP.format(ip_address=ip_address)
+            else:
+                endpoint = self.EP_SIMILARITY_IP.format(ip_address=ip_address)
+
+            response = self._request(endpoint)
+
+            if "ip" not in response:
+                response["ip"] = ip_address
+
+        return response
+
+    def timeline(self, ip_address, field="classification", days=None, granularity=None):
+        """Query IP on the IP TimeSeries API
+
+        :param ip_address: IP address to use in the look-up.
+        :type ip_address: str
+        :param field: field name to use to retrieve timeline information
+        :type field: str
+        :param days: Number of days to show data for
+        :type days: int
+        :param granularity: Granularity of activity date ranges
+        :type granularity: str
+        :return: Context for the IP address.
+        :rtype: dict
+
+
+        """
+        if self.offering == "community":
+            response = {"message": "Similarity lookup not supported with Community offering"}
+        else:
+            LOGGER.debug("Checking IP Sim results for %s...", ip_address)
+            validate_ip(ip_address)
+            if not field:
+                field = "classification"
+            validate_timeline_field_value(field)
+            if days:
+                validate_timeline_days(days)
+            if granularity:
+                validate_timeline_granularity(granularity)
+
+            endpoint = self.EP_TIMELINE_IP.format(ip_address=ip_address)
+            endpoint = endpoint + f"?field={field.lower()}"
+            if days:
+                endpoint = endpoint + f"&days={days}"
+            if granularity:
+                endpoint = endpoint + f"&granularity={granularity}"
+            response = self._request(endpoint)
+
+            if "ip" not in response:
+                response["ip"] = ip_address
+
+        return response
+
+    def timelinedetails(self, ip_address, days=None, cursor=None, limit=100):
+        """Query IP on the IP TimeSeries API
+
+        :param ip_address: IP address to use in the look-up.
+        :type ip_address: str
+        :param cursor: The cursor is a pointer from which to start returning results up to the limit
+        :type cursor: str
+        :param days: Number of days to show data for
+        :type days: int
+        :param limit: The total number of events to return in the response
+        :type limit: str
+        :return: Context for the IP address.
+        :rtype: dict
+
+
+        """
+        if self.offering == "community":
+            response = {"message": "Similarity lookup not supported with Community offering"}
+        else:
+            LOGGER.debug("Checking IP Sim results for %s...", ip_address)
+            validate_ip(ip_address)
+            if days:
+                validate_timeline_days(days)
+
+            endpoint = self.EP_TIMELINE_DETAILS_IP.format(ip_address=ip_address)
+            endpoint = endpoint + f"?limit={limit}"
+            if days:
+                endpoint = endpoint + f"&days={days}"
+            if cursor:
+                endpoint = endpoint + f"&cursor={cursor}"
+            print(endpoint)
             response = self._request(endpoint)
 
             if "ip" not in response: