+++ src/greynoise/cli/subcommand.py
@@ -335,3 +335,27 @@ def similar(
     ip_addresses = get_ip_addresses(context, input_file, ip_address)
     results = [api_client.similar(ip_address=ip_address) for ip_address in ip_addresses]
     return results
+
+
+@ip_lookup_command
+@click.option("-v", "--verbose", count=True, help="Verbose output")
+@click.option("-d", "--days", type=int, help="Number of Days to display")
+@click.option("-F", "--field_name", help="Field name to display data for")
+def timeline(
+    context,
+    api_client,
+    api_key,
+    input_file,
+    output_file,
+    output_format,
+    verbose,
+    ip_address,
+    offering,
+    field_name,
+    days
+):
+    """Query GreyNoise IP to identify Similar IPs."""
+    ip_addresses = get_ip_addresses(context, input_file, ip_address)
+    results = [api_client.timeline(ip_address=ip_address, days=days, field=field_name) for
+               ip_address in ip_addresses]
+    return results